requirements.txt
streamlit
pillow
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
